+++ spec/integration/resource/api300/c7000/web_server_certificate/delete_spec.rb
@@ -1,0 +1,7 @@
+require 'spec_helper'
+
+klass = OneviewSDK::API300::C7000::WebServerCertificate
+RSpec.describe klass, integration: true, type: DELETE, sequence: rseq(klass) do
+  let(:current_client) { $client_300 }
+  include_examples 'WebServerCertificateDeleteExample', 'integration api300 context'
+end