+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Suggested release: v5.0.0
 
+#### New Resources:
+This release adds support support to Oneview Rest API version 200 (only hardware variant C7000), 300 and 500 (both two for the hardware variants C7000 and Synergy) to the new features:
+   - Event
+   - Web Server Certificate
+
 #### Features supported
 This release adds support to OneView Rest API version 500 for the hardware variants C7000 and Synergy to the already existing features:
    - Connection Template